Migration Tool - Adding Events, updating AdminEvents.
