+++ tests/migration-util/src/main/java/org/keycloak/test/migration/MigrateTest.java
@@ -5,17 +5,13 @@ import java.io.BufferedWriter;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.stream.Stream;
-
-import static java.lang.Thread.sleep;
 
 public class MigrateTest {
 
@@ -26,8 +22,9 @@ public class MigrateTest {
             ChangePackageRewrite.class,
             RenameImportsRewrite.class,
             UpdateAssertsRewrite.class,
-            AddManagedResourcesRewrite.class,
+            AddManagedRealmRewrite.class,
             AdminEventAssertRewrite.class,
+            EventAssertRewrite.class,
             BeforeRewrite.class,
             AfterRewrite.class,
             OAuthClientRewrite.class,
